Add container version
When we make a change to the spring boot app, the tast definition is not
updated, and the new image is not deployed.
Currently we always use the "latest" image version, I believe because of
this Terraform does not detect any changes to the task definition.  By
using the version rather than the latest, if there is a change it should
be detected.

This change adds back in the container version, used in a similar way to
OPG.

linted-by: git-pre-commit-lint.sh v1.8.4

```
Binaries required for validation:
	yamllint:	OK      ✔
	tflint:		OK      ✔

Checked files:
	.circleci/config.yml
	terraform/modules/app-cluster/
	terraform/

Unchecked or partially validated files:
	terraform/modules/app-cluster/templates/hello_world.json.tpl
```

linted-by: git-pre-commit-lint.sh v1.8.4

```
Binaries required for validation:
	yamllint:	OK      ✔
	tflint:		OK      ✔

Checked files:
	.circleci/config.yml
	terraform/modules/app-cluster/
```

linted-by: git-pre-commit-lint.sh v1.8.4

```
Binaries required for validation:
	yamllint:	OK      ✔

Checked files:
	.circleci/config.yml
```

linted-by: git-pre-commit-lint.sh v1.8.4

```

Unchecked or partially validated files:
	terraform/modules/app-cluster/templates/hello_world.json.tpl
```

linted-by: git-pre-commit-lint.sh v1.8.4

```
Binaries required for validation:
	yamllint:	OK      ✔

Checked files:
	.circleci/config.yml
```

linted-by: git-pre-commit-lint.sh v1.8.4

```

Unchecked or partially validated files:
	src/main/java/hello/HelloController.java
	src/test/java/hello/HelloControllerTest.java
```

linted-by: git-pre-commit-lint.sh v1.8.4

```
Binaries required for validation:
	tflint:		OK      ✔

Checked files:
	terraform/
```

linted-by: git-pre-commit-lint.sh v1.8.4

```
Binaries required for validation:
	yamllint:	OK      ✔

Checked files:
	.circleci/config.yml
```